requirements.txt
fastapi
typing
strawberry-graphql[debug-server]
starlette
"uvicorn[standard]"
